+++ TravelList/TravelListTableViewController.swift
@@ -27,15 +27,8 @@ class TravelListTableViewController: UITableViewController {
         
         logoView = UIImageView(frame: CGRectMake(0, 0, 40, 40))
         logoView.image = UIImage(named: "logoTravel")?.imageWithRenderingMode(.AlwaysTemplate)
-        logoView.frame.origin.x = (self.view.frame.size.width - logoView.frame.size.width) / 2
-        logoView.frame.origin.y = 20
         logoView.tintColor = toColor("#F9F9F9")
-        
-        // Add the logo view to the navigation controller.
-        self.navigationController?.view.addSubview(logoView)
-        
-        // Bring the logo view to the front.
-        self.navigationController?.view.bringSubviewToFront(logoView)
+        self.navigationItem.titleView = logoView
         
         loadInitialData()
     }
@@ -116,6 +109,5 @@ class TravelListTableViewController: UITableViewController {
         
     }
 }
-